* Don't truncate the result in VecToAngles().
* See bug #3210 (Accuracy system with laser guns)

+++ src/shared/mathlib.cpp
@@ -965,25 +965,25 @@ void VecToAngles (const vec3_t value1, vec3_t angles)
 
 	/* only check the first two values for being zero */
 	if (Vector2Empty(value1)) {
-		yaw = 0;
+		yaw = 0.0f;
 		if (value1[2] > 0)
-			pitch = 90.0;
+			pitch = 90.0f;
 		else
-			pitch = 270.0;
+			pitch = 270.0f;
 	} else {
 		const float forward = sqrt(value1[0] * value1[0] + value1[1] * value1[1]);
 		if (!EQUAL(value1[0], 0.0))
-			yaw = (int) (atan2(value1[1], value1[0]) * todeg);
-		else if (value1[1] > 0)
-			yaw = 90.0;
+			yaw = atan2(value1[1], value1[0]) * todeg;
+		else if (value1[1] > 0.0f)
+			yaw = 90.0f;
 		else
-			yaw = -90.0;
+			yaw = -90.0f;
 		if (yaw < 0.0)
-			yaw += 360.0;
+			yaw += 360.0f;
 
-		pitch = (int) (atan2(value1[2], forward) * todeg);
-		if (pitch < 0)
-			pitch += 360.0;
+		pitch = atan2(value1[2], forward) * todeg;
+		if (pitch < 0.0f)
+			pitch += 360.0f;
 	}
 
 	/* up and down */
@@ -991,9 +991,8 @@ void VecToAngles (const vec3_t value1, vec3_t angles)
 	/* left and right */
 	angles[YAW] = yaw;
 	/* tilt left and right */
-	angles[ROLL] = 0;
+	angles[ROLL] = 0.0f;
 }
-
 
 /**
  * @brief Checks whether i is power of two value
